[KARAF-3147] Remove the KarafMBeanServerBuilder and refactore the guard interception
