+++ assemblies/apache-karaf/pom.xml
@@ -530,12 +530,6 @@
                                     <outputDirectory>target/dependencies/lib</outputDirectory>
                                     <destFileName>karaf-jaas-boot.jar</destFileName>
                                 </artifactItem>
-                                <artifactItem>
-                                    <groupId>org.apache.karaf.management</groupId>
-                                    <artifactId>org.apache.karaf.management.boot</artifactId>
-                                    <outputDirectory>target/dependencies/lib</outputDirectory>
-                                    <destFileName>karaf-jmx-boot.jar</destFileName>
-                                </artifactItem>                                
                             </artifactItems>
                         </configuration>
                     </execution>
