+++ itests/src/test/java/org/apache/karaf/itests/JMXSecurityTest.java
@@ -70,7 +70,6 @@ public class JMXSecurityTest extends KarafTestSupport {
 
         // Add some extra options used by this test...
         options.addAll(Arrays.asList(
-            new VMOption("-Djavax.management.builder.initial=org.apache.karaf.management.boot.KarafMBeanServerBuilder"),
             editConfigurationFilePut("etc/jmx.acl.org.apache.karaf.services.cfg", "getServices(boolean)", "viewer"),
             editConfigurationFilePut("etc/jmx.acl.org.apache.karaf.services.cfg", "getServices(long)", "manager"),
             editConfigurationFilePut("etc/jmx.acl.org.apache.karaf.services.cfg", "getServices(long,boolean)", "admin")));
