--- management/boot/src/main/resources/OSGI-INF/bundle.info
@@ -1,17 +1,0 @@
-\u001B[1mSYNOPSIS\u001B[0m
-    ${project.name}
-
-    ${project.description}
-
-    Maven URL:
-        \u001B[33mmvn:${project.groupId}/${project.artifactId}/${project.version}\u001B[0m
-
-\u001B[1mDESCRIPTION\u001B[0m
-    This bundle provides JMX classes loaded during Karaf boot process.
-
-    In particular it provides the KarafMBeanServerBuilder.
-
-    The KarafMBeanServerBuilder builds a special MBeanServer that allows role-based authorization
-    for JMX access.
-    This class must be loadable from at startup by the JVM so this module is added to classpath, the
-    the boot delegation class path and made available via the system bundle.
