+++ shell/wrapper/src/main/resources/org/apache/karaf/shell/wrapper/windows/karaf-wrapper.conf
@@ -31,8 +31,7 @@ wrapper.java.mainclass=org.apache.karaf.shell.wrapper.Main
 wrapper.java.classpath.1=%KARAF_BASE%/lib/karaf-wrapper.jar
 wrapper.java.classpath.2=%KARAF_HOME%/lib/karaf.jar
 wrapper.java.classpath.3=%KARAF_HOME%/lib/karaf-jaas-boot.jar
-wrapper.java.classpath.4=%KARAF_HOME%/lib/karaf-jmx-boot.jar
-wrapper.java.classpath.5=%KARAF_BASE%/lib/karaf-wrapper-main.jar
+wrapper.java.classpath.4=%KARAF_BASE%/lib/karaf-wrapper-main.jar
 wrapper.java.library.path.1=%KARAF_BASE%/lib/
 
 # Application Parameters.  Add parameters as needed starting from 1
@@ -45,11 +44,10 @@ wrapper.java.additional.2=-Dkaraf.base="%KARAF_BASE%"
 wrapper.java.additional.3=-Dkaraf.data="%KARAF_DATA%"
 wrapper.java.additional.4=-Dkaraf.etc="%KARAF_ETC%"
 wrapper.java.additional.5=-Dcom.sun.management.jmxremote
-wrapper.java.additional.6=-Djavax.management.builder.initial=org.apache.karaf.management.boot.KarafMBeanServerBuilder
-wrapper.java.additional.7=-Dkaraf.startLocalConsole=false
-wrapper.java.additional.8=-Dkaraf.startRemoteShell=true
-wrapper.java.additional.9=-Djava.endorsed.dirs="%JAVA_HOME%/jre/lib/endorsed;%JAVA_HOME%/lib/endorsed;%KARAF_HOME%/lib/endorsed"
-wrapper.java.additional.10=-Djava.ext.dirs="%JAVA_HOME%/jre/lib/ext;%JAVA_HOME%/lib/ext;%KARAF_HOME%/lib/ext"
+wrapper.java.additional.6=-Dkaraf.startLocalConsole=false
+wrapper.java.additional.7=-Dkaraf.startRemoteShell=true
+wrapper.java.additional.8=-Djava.endorsed.dirs="%JAVA_HOME%/jre/lib/endorsed;%JAVA_HOME%/lib/endorsed;%KARAF_HOME%/lib/endorsed"
+wrapper.java.additional.9=-Djava.ext.dirs="%JAVA_HOME%/jre/lib/ext;%JAVA_HOME%/lib/ext;%KARAF_HOME%/lib/ext"
 
 # Uncomment to enable jmx
 #wrapper.java.additional.n=-Dcom.sun.management.jmxremote.port=1616
